+++ compiler-core/src/metadata/tests.rs
@@ -75,6 +75,7 @@ fn bit_array_segment_option_module(option: TypedConstantBitArraySegmentOption) -
             value: Box::new(Constant::Int {
                 location: Default::default(),
                 value: "1".into(),
+                int_value: 1.into(),
             }),
             options: vec![option],
             type_: type_::int(),
@@ -834,6 +835,7 @@ fn constant_int() {
     let module = constant_module(Constant::Int {
         location: Default::default(),
         value: "100".into(),
+        int_value: 100.into(),
     });
 
     assert_eq!(roundtrip(&module), module);
@@ -867,6 +869,7 @@ fn constant_tuple() {
             Constant::Int {
                 location: Default::default(),
                 value: "1".into(),
+                int_value: 1.into(),
             },
             Constant::Float {
                 location: Default::default(),
@@ -878,6 +881,7 @@ fn constant_tuple() {
                     Constant::Int {
                         location: Default::default(),
                         value: "1".into(),
+                        int_value: 1.into(),
                     },
                     Constant::Float {
                         location: Default::default(),
@@ -900,14 +904,17 @@ fn constant_list() {
             Constant::Int {
                 location: Default::default(),
                 value: "1".into(),
+                int_value: 1.into(),
             },
             Constant::Int {
                 location: Default::default(),
                 value: "2".into(),
+                int_value: 2.into(),
             },
             Constant::Int {
                 location: Default::default(),
                 value: "3".into(),
+                int_value: 3.into(),
             },
         ],
     });
@@ -938,6 +945,7 @@ fn constant_record() {
                 value: Constant::Int {
                     location: Default::default(),
                     value: "1".into(),
+                    int_value: 1.into(),
                 },
             },
         ],
@@ -954,6 +962,7 @@ fn constant_var() {
     let one_original = Constant::Int {
         location: Default::default(),
         value: "1".into(),
+        int_value: 1.into(),
     };
 
     let one = Constant::Var {
@@ -1084,6 +1093,7 @@ fn constant_bit_array_size() {
         value: Box::new(Constant::Int {
             location: Default::default(),
             value: "1".into(),
+            int_value: 1.into(),
         }),
         short_form: false,
     });
@@ -1097,6 +1107,7 @@ fn constant_bit_array_size_short_form() {
         value: Box::new(Constant::Int {
             location: Default::default(),
             value: "1".into(),
+            int_value: 1.into(),
         }),
         short_form: true,
     });